Add files via upload
Note that this is a very basic example and wouldn't work for your case as-is. The image_recognition_module and the is_tobacco function don't exist - they're placeholders for a complex image recognition algorithm that could potentially identify tobacco in an image, which would be highly challenging to create. Even with a well-trained model, the task would still be very difficult due to the small size and lack of distinct features of discarded tobacco.